adds elastic search test into gh action pipeline with a custom integration test, uses mocks for testing elastic locally, some other cleanup
