+++ .github/workflows/pull-request.yml
@@ -6,6 +6,7 @@ on:
 jobs:
   build:
     runs-on: ubuntu-latest
+
     strategy:
       matrix:
         operating-system: [ ubuntu-20.04 ]
@@ -22,6 +23,19 @@ jobs:
           php-version: ${{ matrix.php-versions }}
           extensions: mbstring, intl, xdebug
 
+      - name: Configure sysctl limits
+        run: |
+          sudo swapoff -a
+          sudo sysctl -w vm.swappiness=1
+          sudo sysctl -w fs.file-max=262144
+          sudo sysctl -w vm.max_map_count=262144
+
+      - name: Runs Elasticsearch
+        uses: elastic/elastic-github-actions/elasticsearch@master
+        with:
+          stack-version: 6.2.4
+          security-enabled: false
+
       - name: PHP Version
         run: php -v
 
@@ -35,6 +49,7 @@ jobs:
       - name: Test Suite
         run: |
           composer test
+          vendor/bin/phpunit tests/TestCase/Persister/ElasticSearchPersisterIntegrationTest.php
           echo ${{ matrix.php-versions }}
 
   #
@@ -58,6 +73,19 @@ jobs:
       - name: PHP Version
         run: php -v
 
+      - name: Configure sysctl limits
+        run: |
+          sudo swapoff -a
+          sudo sysctl -w vm.swappiness=1
+          sudo sysctl -w fs.file-max=262144
+          sudo sysctl -w vm.max_map_count=262144
+
+      - name: Runs Elasticsearch
+        uses: elastic/elastic-github-actions/elasticsearch@master
+        with:
+          stack-version: 6.2.4
+          security-enabled: false
+
       - name: CakePHP ${{matrix.version}} Compatability
         run: |
           composer self-update
@@ -65,6 +93,7 @@ jobs:
           composer require cakephp/cakephp:${{matrix.version}} --no-update
           composer install --prefer-dist --no-progress
           composer test
+          vendor/bin/phpunit tests/TestCase/Persister/ElasticSearchPersisterIntegrationTest.php
 
   compatibility_cakephp4:
     runs-on: ubuntu-latest
@@ -84,6 +113,19 @@ jobs:
       - name: PHP Version
         run: php -v
 
+      - name: Configure sysctl limits
+        run: |
+          sudo swapoff -a
+          sudo sysctl -w vm.swappiness=1
+          sudo sysctl -w fs.file-max=262144
+          sudo sysctl -w vm.max_map_count=262144
+
+      - name: Runs Elasticsearch
+        uses: elastic/elastic-github-actions/elasticsearch@master
+        with:
+          stack-version: 6.2.4
+          security-enabled: false
+
       - name: CakePHP ${{matrix.version}} Compatability
         run: |
           composer self-update
@@ -91,3 +133,4 @@ jobs:
           composer require cakephp/cakephp:${{matrix.version}} --no-update
           composer install --prefer-dist --no-progress
           composer test
+          vendor/bin/phpunit tests/TestCase/Persister/ElasticSearchPersisterIntegrationTest.php
